% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellAlleleCounts-class.R
\name{CellAlleleCounts_from_dataframe}
\alias{CellAlleleCounts_from_dataframe}
\title{Convert a dataframe produced by sclantern-nf into CellAlleleCounts}
\usage{
CellAlleleCounts_from_dataframe(df_counts, min_cells_per_feature = 2)
}
\arguments{
\item{df_counts}{A long dataframe or data.table, as produced by
sclantern-nf. The table should have columns "cell", "umi",
"allele", "locus", "count", corresponding respectively to the
cell barcode, UMI barcode, allele, locus, reads.}

\item{cell_data}{Optional DataFrame of cell metadata. Rownames must
match column names of the matrices. If NULL, an empty DataFrame
is created.}
}
\value{
A \code{CellAlleleCounts} object.
}
\description{
Convert a dataframe produced by sclantern-nf into CellAlleleCounts
}
