% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeF2dist.R
\name{computeF2dist}
\alias{computeF2dist}
\title{Compute F2-distance from (imputed) allele frequencies}
\usage{
computeF2dist(freqs)
}
\arguments{
\item{freqs}{A list of per-locus allele frequency matrices (A_l x N) as
returned by \code{predict()}, or a precomputed N x N inner
product matrix.}
}
\value{
A matrix of the F2 distances between every pair of cells.
}
\description{
Compute F2-distance from (imputed) allele frequencies
}
